requirements.txt
streamlit==1.32.0
langchain==0.1.13
langchain-community==0.0.28
chromadb==0.4.24
python-dotenv==1.0.1
pypdf==3.17.4
docx2txt==0.8
ollama==0.1.6
langchain-ollama==0.1.0